+++ Content.Client/Access/UI/AgentIDCardBoundUserInterface.cs
@@ -1,6 +1,5 @@
 using Content.Shared.Access.Systems;
 using Content.Shared.StatusIcon;
-using Robust.Client.GameObjects;
 using Robust.Client.UserInterface;
 using Robust.Shared.Prototypes;
