+++ Content.Server/Access/Systems/AgentIDCardSystem.cs
@@ -1,42 +1,24 @@
-using Content.Server.Access.Components;
-using Content.Server.Popups;
-using Content.Shared.UserInterface;
 using Content.Shared.Access.Components;
-using Content.Shared.Access.Systems;
-using Content.Shared.Interaction;
-using Content.Shared.StatusIcon;
-using Robust.Server.GameObjects;
 using Robust.Shared.Prototypes;
-using Content.Shared.Roles;
-using System.Diagnostics.CodeAnalysis;
 using Content.Server.Clothing.Systems;
 using Content.Server.Implants;
+using Content.Shared.Access.Systems;
 using Content.Shared.Implants;
 using Content.Shared.Inventory;
-using Content.Shared.Lock;
 using Content.Shared.PDA;
 
 namespace Content.Server.Access.Systems
 {
     public sealed class AgentIDCardSystem : SharedAgentIdCardSystem
     {
-        [Dependency] private readonly PopupSystem _popupSystem = default!;
         [Dependency] private readonly IdCardSystem _cardSystem = default!;
-        [Dependency] private readonly UserInterfaceSystem _uiSystem = default!;
         [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
         [Dependency] private readonly ChameleonClothingSystem _chameleon = default!;
         [Dependency] private readonly ChameleonControllerSystem _chamController = default!;
-        [Dependency] private readonly LockSystem _lock = default!;
 
         public override void Initialize()
         {
             base.Initialize();
-            SubscribeLocalEvent<AgentIDCardComponent, AfterInteractEvent>(OnAfterInteract);
-            // BUI
-            SubscribeLocalEvent<AgentIDCardComponent, AfterActivatableUIOpenEvent>(AfterUIOpen);
-            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardNameChangedMessage>(OnNameChanged);
-            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardJobChangedMessage>(OnJobChanged);
-            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardJobIconChangedMessage>(OnJobIconChanged);
             SubscribeLocalEvent<AgentIDCardComponent, InventoryRelayedEvent<ChameleonControllerOutfitSelectedEvent>>(OnChameleonControllerOutfitChangedItem);
         }
 
@@ -77,81 +59,6 @@ namespace Content.Server.Access.Systems
                 return;
 
             _chameleon.SetSelectedPrototype(ent, comp.IdCard);
-        }
-
-        private void OnAfterInteract(EntityUid uid, AgentIDCardComponent component, AfterInteractEvent args)
-        {
-            if (args.Target == null || !args.CanReach || _lock.IsLocked(uid) ||
-                !TryComp<AccessComponent>(args.Target, out var targetAccess) || !HasComp<IdCardComponent>(args.Target))
-                return;
-
-            if (!TryComp<AccessComponent>(uid, out var access) || !HasComp<IdCardComponent>(uid))
-                return;
-
-            var beforeLength = access.Tags.Count;
-            access.Tags.UnionWith(targetAccess.Tags);
-            var addedLength = access.Tags.Count - beforeLength;
-
-            _popupSystem.PopupEntity(Loc.GetString("agent-id-new", ("number", addedLength), ("card", args.Target)), args.Target.Value, args.User);
-            if (addedLength > 0)
-                Dirty(uid, access);
-        }
-
-        private void AfterUIOpen(EntityUid uid, AgentIDCardComponent component, AfterActivatableUIOpenEvent args)
-        {
-            if (!_uiSystem.HasUi(uid, AgentIDCardUiKey.Key))
-                return;
-
-            if (!TryComp<IdCardComponent>(uid, out var idCard))
-                return;
-
-            var state = new AgentIDCardBoundUserInterfaceState(idCard.FullName ?? "", idCard.LocalizedJobTitle ?? "", idCard.JobIcon);
-            _uiSystem.SetUiState(uid, AgentIDCardUiKey.Key, state);
-        }
-
-        private void OnJobChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardJobChangedMessage args)
-        {
-            if (!TryComp<IdCardComponent>(uid, out var idCard))
-                return;
-
-            _cardSystem.TryChangeJobTitle(uid, args.Job, idCard);
-        }
-
-        private void OnNameChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardNameChangedMessage args)
-        {
-            if (!TryComp<IdCardComponent>(uid, out var idCard))
-                return;
-
-            _cardSystem.TryChangeFullName(uid, args.Name, idCard);
-        }
-
-        private void OnJobIconChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardJobIconChangedMessage args)
-        {
-            if (!TryComp<IdCardComponent>(uid, out var idCard))
-                return;
-
-            if (!_prototypeManager.TryIndex(args.JobIconId, out var jobIcon))
-                return;
-
-            _cardSystem.TryChangeJobIcon(uid, jobIcon, idCard);
-
-            if (TryFindJobProtoFromIcon(jobIcon, out var job))
-                _cardSystem.TryChangeJobDepartment(uid, job, idCard);
-        }
-
-        private bool TryFindJobProtoFromIcon(JobIconPrototype jobIcon, [NotNullWhen(true)] out JobPrototype? job)
-        {
-            foreach (var jobPrototype in _prototypeManager.EnumeratePrototypes<JobPrototype>())
-            {
-                if (jobPrototype.Icon == jobIcon.ID)
-                {
-                    job = jobPrototype;
-                    return true;
-                }
-            }
-
-            job = null;
-            return false;
         }
     }
 }
