+++ Content.Shared/Access/Components/AgentIDCardComponent.cs
@@ -1,4 +1,4 @@
-namespace Content.Server.Access.Components;
+namespace Content.Shared.Access.Components;
 
 /// <summary>
 ///     Allows an ID card to copy accesses from other IDs and to change the name, job title and job icon via an interface.
