+++ Content.Shared/Access/SharedAgentIDCardSystem.cs
@@ -1,9 +1,111 @@
+using System.Diagnostics.CodeAnalysis;
+using Content.Shared.Access.Components;
+using Content.Shared.Interaction;
+using Content.Shared.Lock;
+using Content.Shared.Popups;
+using Content.Shared.Roles;
 using Content.Shared.StatusIcon;
+using Content.Shared.UserInterface;
 using Robust.Shared.Prototypes;
 using Robust.Shared.Serialization;
 
 namespace Content.Shared.Access.Systems
 {
+    public sealed class AgentIDCardSystem : SharedAgentIdCardSystem
+    {
+        [Dependency] private readonly SharedPopupSystem _popupSystem = default!;
+        [Dependency] private readonly SharedIdCardSystem _cardSystem = default!;
+        [Dependency] private readonly SharedUserInterfaceSystem _uiSystem = default!;
+        [Dependency] private readonly IPrototypeManager _prototypeManager = default!;
+        [Dependency] private readonly LockSystem _lock = default!;
+
+        public override void Initialize()
+        {
+            base.Initialize();
+            SubscribeLocalEvent<AgentIDCardComponent, AfterInteractEvent>(OnAfterInteract);
+            // BUI
+            SubscribeLocalEvent<AgentIDCardComponent, AfterActivatableUIOpenEvent>(AfterUIOpen);
+            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardNameChangedMessage>(OnNameChanged);
+            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardJobChangedMessage>(OnJobChanged);
+            SubscribeLocalEvent<AgentIDCardComponent, AgentIDCardJobIconChangedMessage>(OnJobIconChanged);
+        }
+
+        private void OnAfterInteract(EntityUid uid, AgentIDCardComponent component, AfterInteractEvent args)
+        {
+            if (args.Target == null || !args.CanReach || _lock.IsLocked(uid) ||
+                !TryComp<AccessComponent>(args.Target, out var targetAccess) || !HasComp<IdCardComponent>(args.Target))
+                return;
+
+            if (!TryComp<AccessComponent>(uid, out var access) || !HasComp<IdCardComponent>(uid))
+                return;
+
+            var beforeLength = access.Tags.Count;
+            access.Tags.UnionWith(targetAccess.Tags);
+            var addedLength = access.Tags.Count - beforeLength;
+
+            _popupSystem.PopupPredicted(Loc.GetString("agent-id-new", ("number", addedLength), ("card", args.Target)), args.Target.Value, args.User);
+            if (addedLength > 0)
+                Dirty(uid, access);
+        }
+
+        private void AfterUIOpen(EntityUid uid, AgentIDCardComponent component, AfterActivatableUIOpenEvent args)
+        {
+            if (!_uiSystem.HasUi(uid, AgentIDCardUiKey.Key))
+                return;
+
+            if (!TryComp<IdCardComponent>(uid, out var idCard))
+                return;
+
+            var state = new AgentIDCardBoundUserInterfaceState(idCard.FullName ?? "", idCard.LocalizedJobTitle ?? "", idCard.JobIcon);
+            _uiSystem.SetUiState(uid, AgentIDCardUiKey.Key, state);
+        }
+
+        private void OnJobChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardJobChangedMessage args)
+        {
+            if (!TryComp<IdCardComponent>(uid, out var idCard))
+                return;
+
+            _cardSystem.TryChangeJobTitle(uid, args.Job, idCard);
+        }
+
+        private void OnNameChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardNameChangedMessage args)
+        {
+            if (!TryComp<IdCardComponent>(uid, out var idCard))
+                return;
+
+            _cardSystem.TryChangeFullName(uid, args.Name, idCard);
+        }
+
+        private void OnJobIconChanged(EntityUid uid, AgentIDCardComponent comp, AgentIDCardJobIconChangedMessage args)
+        {
+            if (!TryComp<IdCardComponent>(uid, out var idCard))
+                return;
+
+            if (!_prototypeManager.TryIndex(args.JobIconId, out var jobIcon))
+                return;
+
+            _cardSystem.TryChangeJobIcon(uid, jobIcon, idCard);
+
+            if (TryFindJobProtoFromIcon(jobIcon, out var job))
+                _cardSystem.TryChangeJobDepartment(uid, job, idCard);
+        }
+
+        private bool TryFindJobProtoFromIcon(JobIconPrototype jobIcon, [NotNullWhen(true)] out JobPrototype? job)
+        {
+            foreach (var jobPrototype in _prototypeManager.EnumeratePrototypes<JobPrototype>())
+            {
+                if (jobPrototype.Icon == jobIcon.ID)
+                {
+                    job = jobPrototype;
+                    return true;
+                }
+            }
+
+            job = null;
+            return false;
+        }
+    }
+
     public abstract class SharedAgentIdCardSystem : EntitySystem
     {
         // Just for friending for now
